Funcionalidades nuevas, cambiar recibo, nombre con destino, agrandar comprobante

+++ FrontendGPS/src/pages/pagoPage.jsx
@@ -1,9 +1,11 @@
 import { useState, useEffect } from 'react';
 import axios from 'axios';
 import jsPDF from 'jspdf';
+import emailjs from '@emailjs/browser';
 import CouponList from '../components/couponComponents/couponList';
 import CouponForm from '../components/couponComponents/couponForm';
 import PublicationSelect from '../components/couponComponents/PublicationSelect';
+import EditCouponModal from '../components/couponComponents/EditCouponModal';
 
 function PagoPage() {
     const [coupon, setCoupon] = useState(null);
@@ -17,6 +19,9 @@ function PagoPage() {
     const [searchRut, setSearchRut] = useState('');
     const [adminCoupons, setAdminCoupons] = useState([]);
     const [errorMessage, setErrorMessage] = useState('');
+    const [showEditModal, setShowEditModal] = useState(false);
+    const [selectedCoupon, setSelectedCoupon] = useState(null);
+
 
     useEffect(() => {
         if (typeof window !== 'undefined' && window.localStorage) {
@@ -65,6 +70,25 @@ function PagoPage() {
         }
     };
 
+    const sendEmailToAdmin = async (subject, message) => {
+        try {
+            await emailjs.send(
+                import.meta.env.VITE_EMAILJS_SERVICE_ID,
+                import.meta.env.VITE_EMAILJS_TEMPLATE_ID,
+                {
+                    subject,
+                    message,
+                    to_email: import.meta.env.VITE_ADMIN_EMAIL,
+                },
+                import.meta.env.VITE_EMAILJS_PUBLIC_KEY
+            );
+            console.log('Correo enviado con éxito');
+        } catch (error) {
+            console.error('Error al enviar el correo:', error);
+        }
+    };
+
+
     const handleGenerateCoupon = async () => {
         if (!selectedPublication) {
             alert('Por favor, selecciona una publicación.');
@@ -77,24 +101,24 @@ function PagoPage() {
         dueDate.setDate(dueDate.getDate() + 3);
         const dueDateString = dueDate.toISOString();
 
-        let parsedUbicacionDescarga;
+        let interpUbicacionDescarga;
         try {
-            parsedUbicacionDescarga = JSON.parse(ubicacionDescarga);
+            interpUbicacionDescarga = JSON.parse(ubicacionDescarga);
         } catch (e) {
-            console.error("Error al parsear la ubicación de descarga:", e);
+            console.error("Error al interpretar la ubicación de descarga:", e);
             alert("Ubicación de descarga no válida");
             return;
         }
 
-        if (parsedUbicacionDescarga && parsedUbicacionDescarga.lat && parsedUbicacionDescarga.lng) {
-            console.log("Ubicación de descarga válida:", parsedUbicacionDescarga);
+        if (interpUbicacionDescarga && interpUbicacionDescarga.lat && interpUbicacionDescarga.lng) {
+            console.log("Ubicación de descarga válida:", interpUbicacionDescarga);
         } else {
-            console.error("Ubicación de descarga no válida:", parsedUbicacionDescarga);
+            console.error("Ubicación de descarga no válida:", interpUbicacionDescarga);
             alert("Ubicación de descarga no válida");
             return;
         }
 
-        const address = await fetchAddress(parsedUbicacionDescarga.lat, parsedUbicacionDescarga.lng);
+        const address = await fetchAddress(interpUbicacionDescarga.lat, interpUbicacionDescarga.lng);
         console.log("Dirección obtenida:", address);
 
         const data = {
@@ -158,10 +182,17 @@ function PagoPage() {
             doc.text('Este es un cupón de pago generado automáticamente.', 105, 130, null, null, 'center');
 
             doc.save(`Cupon_${data.userId}.pdf`);
+
+            // Enviar correo al admin
+            await sendEmailToAdmin(
+                'Nuevo Cupón Generado',
+                `El usuario ${userId} ha generado un nuevo cupón con ID ${newCoupon._id}.`
+            );
         } catch (error) {
             console.error(`Error al generar el cupón: ${error}`);
         }
     };
+
 
     useEffect(() => {
         const fetchCoupons = async () => {
@@ -185,35 +216,38 @@ function PagoPage() {
     };
 
     const handleDeleteCoupon = async (couponId) => {
-        try {
-            const response = await axios.delete(`http://localhost:3000/cupon/${couponId}`);
-            if (response.status === 200) {
-                // Recarga los cupones después de eliminar uno
-                setCoupons(coupons.filter((coupon) => coupon._id !== couponId));
-            } else {
-                console.error('Error al eliminar el cupón');
+        if (window.confirm("¿Está seguro que desea eliminar este cupón?")) {
+            try {
+                const response = await axios.delete(`http://localhost:3000/cupon/${couponId}`);
+                if (response.status === 200) {
+                    // Actualiza los cupones para eliminar el cupón eliminado
+                    setCoupons(coupons.filter((coupon) => coupon._id !== couponId));
+                    setAdminCoupons(adminCoupons.filter((coupon) => coupon._id !== couponId));
+                } else {
+                    console.error('Error al eliminar el cupón');
+                }
+            } catch (error) {
+                console.error('Error al eliminar el cupón', error);
             }
-        } catch (error) {
-            console.error('Error al eliminar el cupón', error);
         }
     };
 
-    const handleFileChange = (event) => {
+    const handleFileChange = (event, couponId) => {
         const file = event.target.files[0];
         if (file && (file.type === 'image/png' || file.type === 'image/jpeg')) {
-            setSelectedFile(file);
+            setSelectedFile({ file, couponId });
         } else {
             alert('Por favor, sube un archivo .png o .jpg');
         }
     };
 
     const handleUploadFile = async (couponId) => {
-        if (!selectedFile) {
-            alert('No hay archivo seleccionado');
+        if (!selectedFile || selectedFile.couponId !== couponId) {
+            alert('No hay archivo seleccionado o el archivo seleccionado no coincide con el cupón.');
             return;
         }
         const formData = new FormData();
-        formData.append('file', selectedFile);
+        formData.append('file', selectedFile.file);
 
         try {
             const response = await axios.post(`http://localhost:3000/cupon/upload/${couponId}`, formData, {
@@ -224,9 +258,26 @@ function PagoPage() {
             if (response.status === 200) {
                 setCoupons(coupons.map(c => c._id === couponId ? { ...c, receipt: response.data.receipt } : c));
                 setSelectedFile(null);
+
+
+                await sendEmailToAdmin(
+                    'Comprobante Subido',
+                    `El usuario ${userId} ha subido un comprobante para el cupón con ID ${couponId}.`
+                );
             }
         } catch (error) {
             console.error('Error al subir el archivo:', error);
+        }
+    };
+
+    const handleDeleteReceipt = async (couponId) => {
+        try {
+            const response = await axios.delete(`http://localhost:3000/cupon/${couponId}/delete-image/`);
+            if (response.status === 200) {
+                setCoupons(coupons.map(c => c._id === couponId ? { ...c, receipt: null } : c));
+            }
+        } catch (error) {
+            console.error('Error al eliminar el comprobante:', error);
         }
     };
 
@@ -249,6 +300,23 @@ function PagoPage() {
                 setErrorMessage('Error al buscar cupones por RUT');
             }
             setAdminCoupons([]);
+        }
+    };
+
+    const handleUpdateCouponStatus = async () => {
+        const newStatus = !selectedCoupon.isPaid;
+        try {
+            const response = await axios.put(`http://localhost:3000/cupon/${selectedCoupon._id}`, {
+                isPaid: newStatus
+            });
+            if (response.status === 200) {
+                setCoupons(coupons.map(c => c._id === selectedCoupon._id ? { ...c, isPaid: newStatus } : c));
+                setAdminCoupons(adminCoupons.map(c => c._id === selectedCoupon._id ? { ...c, isPaid: newStatus } : c));
+                setShowEditModal(false);
+                setSelectedCoupon(null);
+            }
+        } catch (error) {
+            console.error('Error al actualizar el estado del cupón', error);
         }
     };
 
@@ -296,7 +364,10 @@ function PagoPage() {
                     handleFileChange={handleFileChange}
                     handleUploadFile={handleUploadFile}
                     handleDeleteCoupon={handleDeleteCoupon}
-                    userId={userId}
+                    handleDeleteReceipt={handleDeleteReceipt}
+                    setSelectedCoupon={setSelectedCoupon}
+                    setShowEditModal={setShowEditModal}
+                    isAdminView={false}
                 />
             )}
             {userRole === 'admin' && (
@@ -326,11 +397,20 @@ function PagoPage() {
                                 handleFileChange={handleFileChange}
                                 handleUploadFile={handleUploadFile}
                                 handleDeleteCoupon={handleDeleteCoupon}
-                                userId={userId}
+                                setSelectedCoupon={setSelectedCoupon}
+                                setShowEditModal={setShowEditModal}
+                                isAdminView={true}
                             />
                         </div>
                     )}
                 </div>
+            )}
+            {showEditModal && (
+                <EditCouponModal
+                    selectedCoupon={selectedCoupon}
+                    handleUpdateCouponStatus={handleUpdateCouponStatus}
+                    setShowEditModal={setShowEditModal}
+                />
             )}
         </div>
     );
